refactor: upload 함수 분리
