+++ src/store/memberStore.js
@@ -1,6 +1,9 @@
 import { Store } from '../core/store';
-import { db } from '../api/firebase';
-import { collection, doc, getDoc, getDocs } from 'firebase/firestore';
+import { collection, doc, getDoc, getDocs, addDoc } from 'firebase/firestore';
+import { db, storage } from '../api/firebase';
+import { ref, uploadBytes } from 'firebase/storage';
+import { v4 as uuidv4 } from 'uuid';
+
 export const memberStore = new Store({
   members: [],
   loading: false,
@@ -31,4 +34,18 @@ export const getMemberDetail = async (id) => {
     ...res.data(),
     id: res.id,
   };
+};
+
+export const uploadImage = async (fileData) => {
+  const storageRef = ref(storage, uuidv4());
+  await uploadBytes(storageRef, fileData);
+  return storageRef;
+};
+
+export const uploadData = async (data) => {
+  await addDoc(collection(db, 'list'), {
+    name: data.name,
+    email: data.email,
+    photoUrl: data.photoUrl,
+  });
 };
